[MISC] Fix #1673: move terminal.hpp to argument_parser
We decided in core meeting 2019-12-09 to move terminal.hpp into
argument parser.

See core meeting resolution:
* https://github.com/seqan/seqan3/wiki/Core-Meeting-notes#2019-12-09-rrahn

See overview of core changes:
* https://gist.github.com/rrahn/f6b09ad67bebd14bd9c58f77c53e0450
