+++ include/seqan3/argument_parser/argument_parser.hpp
@@ -26,9 +26,9 @@
 #include <seqan3/argument_parser/detail/format_html.hpp>
 #include <seqan3/argument_parser/detail/format_man.hpp>
 #include <seqan3/argument_parser/detail/format_parse.hpp>
+#include <seqan3/argument_parser/detail/terminal.hpp>
 #include <seqan3/argument_parser/detail/version_check.hpp>
 #include <seqan3/core/char_operations/predicate.hpp>
-#include <seqan3/core/detail/terminal.hpp>
 #include <seqan3/core/detail/test_accessor.hpp>
 #include <seqan3/core/detail/to_string.hpp>
 #include <seqan3/io/stream/concept.hpp>
