+++ include/seqan3/argument_parser/detail/format_html.hpp
@@ -15,7 +15,7 @@
 #include <iostream>
 
 #include <seqan3/argument_parser/detail/format_base.hpp>
-#include <seqan3/core/detail/terminal.hpp>
+#include <seqan3/argument_parser/detail/terminal.hpp>
 #include <seqan3/version.hpp>
 
 namespace seqan3::detail
