+++ include/seqan3/argument_parser/detail/version_check.hpp
@@ -20,11 +20,11 @@
 #include <iostream>
 #include <regex>
 
-#include <seqan3/version.hpp>
-#include <seqan3/core/detail/terminal.hpp>
+#include <seqan3/argument_parser/detail/terminal.hpp>
 #include <seqan3/io/detail/misc.hpp>
 #include <seqan3/io/detail/safe_filesystem_entry.hpp>
 #include <seqan3/std/charconv>
+#include <seqan3/version.hpp>
 
 namespace seqan3::detail
 {
