+++ include/seqan3/core/detail/all.hpp
@@ -21,5 +21,4 @@
 #include <seqan3/core/detail/endian.hpp>
 #include <seqan3/core/detail/int_types.hpp>
 #include <seqan3/core/detail/strong_type.hpp>
-#include <seqan3/core/detail/terminal.hpp>
 #include <seqan3/core/detail/type_inspection.hpp>
